Fix bug no. 2: ProfileStorage is now setDirty after reloading params
In test mode, when reloading parameters, ProfileStorage wasn't set to
dirty until the paramaters had been received and the parametersVersion
field was updated. This meant that after first sync, the profile on the
server side was still empty. Setting ProfileStorage dirty straight away
gives us the information a little faster.