Add dummy fields so deserialization doesn't fail
Instead preventing Jackson from failing when it meets unknown
properties, let it fail so that we're informed when the object we wanted
to deserialize to is not the right one. That way we can catch errors
notified by the server in JSON form. The downside to this is we must
define ALL fields that are received from server, in the classes.
Otherwise Jackson fails because it finds unknown properties.