Also updateBEQNotification when pausing BEQActivity
When the final BEQ is done, the `updateBEQNotification()` is called right after
the status of the final BEQ is set to completed (`PageActivity` L417). So the
record in DB hasn't had time to update when `StatusManager` checks if BEQs are
completed (L877). This change makes sure the `updateBEQNotification` is called
one the DB has had time to update.